some tests added

+++ src/main/java/com/scrumtrek/simplestore/FrequentRenterCalculator.java
@@ -4,7 +4,7 @@ package com.scrumtrek.simplestore;
  * Created by user on 11.07.2017.
  */
 public class FrequentRenterCalculator {
-    public float getPoints(PriceCodes priceCode, int days){
+    public int getPoints(PriceCodes priceCode, int days){
         // Add frequent renter points
         int frequentRenterPoints = 1;
